binance-collector/requirements.txt
boto3
pandas
websocket-client
pyarrow
